addressing tons of comments on PR, lab tested
